flexible path to configure, dry-run, documentation
`configure` is no longer required to be into the current directory.  It may be
in the PATH or the full/relative path to configure may be specified.  The path
defaults to the current directory for backward compatibility.

Added the `dry-run` option (-d option)

Updated the output to the `help` action and README.